+++ src/index.js
@@ -1,17 +1,14 @@
 'use strict';
+//required libraries
 var Alexa = require('alexa-sdk');
-var descriptionIntent = require('./intents/getDescriptionIntent')
 var utils = require('./utils')
 var jiraClient = require('./jiraClient')
 
+//intents
+var descriptionIntent = require('./intents/getDescriptionIntent')
+var moveIssueIntent = require('./intents/moveIssueIntent')
+
 var APP_ID = undefined;  // can be replaced with your app ID if publishing
-var MOVE_ISSUE_EN = [
-    "I have moved the issue.",
-    "Jira has moved the issue.",
-    "The issue has been moved.",
-    "I have moved that ticket for you.",
-    "I have moved that ticket."
-]
 var LOAD_ISSUE_EN = [
     "I have loaded issue: ",
     "The issue has been loaded: ",
@@ -21,7 +18,7 @@ var LOAD_ISSUE_EN = [
 ]
 // Test hooks - do not remove!
 exports.GetIssueDescriptionMsg = descriptionIntent.messages;
-exports.MoveIssueMsg = MOVE_ISSUE_EN;
+exports.MoveIssueMsg = moveIssueIntent.messages;
 exports.loadIssueMsg = LOAD_ISSUE_EN;
 var APP_ID_TEST = "mochatest";  // used for mocha tests to prevent warning
 // end Test hooks
@@ -31,7 +28,7 @@ var languageStrings = {
         "translation": {
             "SKILL_NAME": "Jira Skill",  // OPTIONAL change this to a more descriptive name
             "GET_ISSUE_DESCRIPTION": descriptionIntent.messages,
-            "MOVE_ISSUE": MOVE_ISSUE_EN,
+            "MOVE_ISSUE": moveIssueIntent.messages,
             "LOAD_ISSUE": LOAD_ISSUE_EN,
             "HELP_MESSAGE": "You can ask me to do something with a Jira Issue, or, you can say exit... What can I help you with?",
             "HELP_REPROMPT": "What can I help you with?",
@@ -49,7 +46,8 @@ exports.handler = function (event, context, callback) {
     }
     // To enable string internationalization (i18n) features, set a resources object.
     alexa.resources = languageStrings;
-    alexa.registerHandlers(handlers, descriptionIntent.handlers);
+    //Add handlers from ./intents
+    alexa.registerHandlers(handlers, descriptionIntent.handlers, moveIssueIntent.handlers);
     alexa.execute();
 };
 
@@ -92,25 +90,6 @@ var handlers = {
     },
     'MoveIssueIntent': function () {
         this.emit('MoveIssue');
-    },
-    'MoveIssue': function () {
-        var transitionId = this.event.request.intent.slots.TRANSITION_ID.value;
-  
-        if(transitionId){
-            //call the Jira api
-            var moveIssuePromise = jiraClient.moveIssue(transitionId)
-
-            //once the promise is resolved
-            Promise.resolve(moveIssuePromise).then(result => {
-                // Create speech output
-                var speechOutput = utils.getRandomIntroMessage(MOVE_ISSUE_EN);
-                // Use this.t() to get corresponding language data
-                this.emit(':askWithCard', speechOutput, 'You can ask for another task like you did before.', this.t("SKILL_NAME"), speechOutput)
-            })
-        }
-        else{ 
-            this.emit(':askWithCard', 'I did not understand that request', "Please provide a transition ID.", this.t("SKILL_NAME"), 'I did not understand that request. Please provide a valid transition id.');
-        }
     },
     'AMAZON.HelpIntent': function () {
         var speechOutput = this.t("HELP_MESSAGE");
